Fix: support named window in expression.
