+++ src/query/sql/src/planner/binder/select.rs
@@ -33,8 +33,6 @@ use common_ast::ast::SelectTarget;
 use common_ast::ast::SetExpr;
 use common_ast::ast::SetOperator;
 use common_ast::ast::TableReference;
-use common_ast::ast::Window;
-use common_ast::ast::WindowSpec;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_exception::Span;
@@ -115,9 +113,8 @@ impl Binder {
             from_context.all_column_bindings(),
             self.name_resolution_ctx.unquoted_ident_case_sensitive,
         );
-        let (new_stmt, new_order_by) = rewriter.rewrite(stmt, order_by)?;
+        let new_stmt = rewriter.rewrite(stmt)?;
         let stmt = new_stmt.as_ref().unwrap_or(stmt);
-        let order_by = new_order_by.as_deref().unwrap_or(order_by);
 
         // Collect set returning functions
         let set_returning_functions = {
@@ -134,6 +131,10 @@ impl Binder {
         s_expr = self
             .bind_project_set(&mut from_context, &set_returning_functions, s_expr)
             .await?;
+
+        // Try put window definitions into bind context.
+        // This operation should be before `normalize_select_list` because window functions can be used in select list.
+        self.analyze_window_definition(&mut from_context, &stmt.window_list)?;
 
         // Generate a analyzed select list with from context
         let mut select_list = self
@@ -676,7 +677,6 @@ impl Binder {
 struct SelectRewriter<'a> {
     column_binding: &'a [ColumnBinding],
     new_stmt: Option<SelectStmt>,
-    new_order_by: Option<Vec<OrderByExpr>>,
     is_unquoted_ident_case_sensitive: bool,
 }
 
@@ -781,20 +781,14 @@ impl<'a> SelectRewriter<'a> {
         SelectRewriter {
             column_binding,
             new_stmt: None,
-            new_order_by: None,
             is_unquoted_ident_case_sensitive,
         }
     }
 
-    fn rewrite(
-        &mut self,
-        stmt: &SelectStmt,
-        order_by: &[OrderByExpr],
-    ) -> Result<(Option<SelectStmt>, Option<Vec<OrderByExpr>>)> {
-        self.rewrite_window_references(stmt, order_by)?;
+    fn rewrite(&mut self, stmt: &SelectStmt) -> Result<Option<SelectStmt>> {
         self.rewrite_pivot(stmt)?;
         self.rewrite_unpivot(stmt)?;
-        Ok((self.new_stmt.take(), self.new_order_by.take()))
+        Ok(self.new_stmt.take())
     }
 
     fn rewrite_pivot(&mut self, stmt: &SelectStmt) -> Result<()> {
@@ -905,200 +899,5 @@ impl<'a> SelectRewriter<'a> {
             });
         };
         Ok(())
-    }
-
-    fn rewrite_window_references(
-        &mut self,
-        stmt: &SelectStmt,
-        order_by: &[OrderByExpr],
-    ) -> Result<()> {
-        if stmt.window_list.is_none() {
-            return Ok(());
-        }
-        let (window_specs, mut resolved_window_specs) = self.extract_window_definitions(stmt)?;
-
-        let mut window_definitions = HashMap::with_capacity(window_specs.len());
-
-        for (name, spec) in window_specs.iter() {
-            let new_spec = Self::rewrite_inherited_window_spec(
-                spec,
-                &window_specs,
-                &mut resolved_window_specs,
-            )?;
-            window_definitions.insert(name.clone(), new_spec);
-        }
-
-        let mut new_select_list = stmt.select_list.clone();
-        for target in &mut new_select_list {
-            match target {
-                SelectTarget::AliasedExpr { expr, .. } => match expr {
-                    box Expr::FunctionCall { window, .. } => {
-                        if let Some(window) = window {
-                            match window {
-                                Window::WindowReference(reference) => {
-                                    let window_spec = window_definitions
-                                        .get(&reference.window_name.name)
-                                        .ok_or_else(|| {
-                                            ErrorCode::SyntaxException("Window not found")
-                                        })?;
-                                    *window = Window::WindowSpec(WindowSpec {
-                                        existing_window_name: None,
-                                        partition_by: window_spec.partition_by.clone(),
-                                        order_by: window_spec.order_by.clone(),
-                                        window_frame: window_spec.window_frame.clone(),
-                                    });
-                                }
-                                Window::WindowSpec(_) => continue,
-                            }
-                        }
-                    }
-                    _ => continue,
-                },
-                SelectTarget::QualifiedName { .. } => {}
-            }
-        }
-
-        if !order_by.is_empty() {
-            let mut new_order_by = order_by.to_vec();
-            for order in &mut new_order_by {
-                match &mut order.expr {
-                    Expr::FunctionCall { window, .. } => {
-                        if let Some(window) = window {
-                            match window {
-                                Window::WindowReference(reference) => {
-                                    let window_spec = window_definitions
-                                        .get(&reference.window_name.name)
-                                        .ok_or_else(|| {
-                                            ErrorCode::SyntaxException("Window not found")
-                                        })?;
-                                    *window = Window::WindowSpec(WindowSpec {
-                                        existing_window_name: None,
-                                        partition_by: window_spec.partition_by.clone(),
-                                        order_by: window_spec.order_by.clone(),
-                                        window_frame: window_spec.window_frame.clone(),
-                                    });
-                                }
-                                Window::WindowSpec(_) => continue,
-                            }
-                        }
-                    }
-                    _ => continue,
-                }
-            }
-            self.new_order_by = Some(new_order_by);
-        }
-
-        if let Some(ref mut new_stmt) = self.new_stmt {
-            new_stmt.select_list = new_select_list;
-        } else {
-            self.new_stmt = Some(SelectStmt {
-                select_list: new_select_list,
-                ..stmt.clone()
-            });
-        };
-        Ok(())
-    }
-
-    fn extract_window_definitions(
-        &mut self,
-        stmt: &SelectStmt,
-    ) -> Result<(HashMap<String, WindowSpec>, HashMap<String, WindowSpec>)> {
-        let mut window_specs = HashMap::new();
-        let mut resolved_window_specs = HashMap::new();
-        for window in stmt.window_list.as_ref().unwrap() {
-            window_specs.insert(window.name.name.clone(), window.spec.clone());
-            if window.spec.existing_window_name.is_none() {
-                resolved_window_specs.insert(window.name.name.clone(), window.spec.clone());
-            }
-        }
-        if let Some(ref mut new_stmt) = self.new_stmt {
-            new_stmt.window_list = None;
-        } else {
-            self.new_stmt = Some(SelectStmt {
-                window_list: None,
-                ..stmt.clone()
-            });
-        };
-        Ok((window_specs, resolved_window_specs))
-    }
-
-    fn rewrite_inherited_window_spec(
-        window_spec: &WindowSpec,
-        window_list: &HashMap<String, WindowSpec>,
-        resolved_window: &mut HashMap<String, WindowSpec>,
-    ) -> Result<WindowSpec> {
-        if window_spec.existing_window_name.is_some() {
-            let referenced_name = window_spec
-                .existing_window_name
-                .as_ref()
-                .unwrap()
-                .name
-                .clone();
-            // check if spec is resolved first, so that we no need to resolve again.
-            let referenced_window_spec = {
-                resolved_window.get(&referenced_name).unwrap_or(
-                    window_list
-                        .get(&referenced_name)
-                        .ok_or_else(|| ErrorCode::SyntaxException("Referenced window not found"))?,
-                )
-            }
-            .clone();
-
-            let resolved_spec = match referenced_window_spec.existing_window_name.clone() {
-                Some(_) => Self::rewrite_inherited_window_spec(
-                    &referenced_window_spec,
-                    window_list,
-                    resolved_window,
-                )?,
-                None => referenced_window_spec.clone(),
-            };
-
-            // add to resolved.
-            resolved_window.insert(referenced_name, resolved_spec.clone());
-
-            // check semantic
-            if !window_spec.partition_by.is_empty() {
-                return Err(ErrorCode::SemanticError(
-                    "WINDOW specification with named WINDOW reference cannot specify PARTITION BY",
-                ));
-            }
-            if !window_spec.order_by.is_empty() && !resolved_spec.order_by.is_empty() {
-                return Err(ErrorCode::SemanticError(
-                    "Cannot specify ORDER BY if referenced named WINDOW specifies ORDER BY",
-                ));
-            }
-            if resolved_spec.window_frame.is_some() {
-                return Err(ErrorCode::SemanticError(
-                    "Cannot reference named WINDOW containing frame specification",
-                ));
-            }
-
-            // resolve referenced window
-            let mut partition_by = window_spec.partition_by.clone();
-            if !resolved_spec.partition_by.is_empty() {
-                partition_by = resolved_spec.partition_by.clone();
-            }
-
-            let mut order_by = window_spec.order_by.clone();
-            if order_by.is_empty() && !resolved_spec.order_by.is_empty() {
-                order_by = resolved_spec.order_by.clone();
-            }
-
-            let mut window_frame = window_spec.window_frame.clone();
-            if window_frame.is_none() && resolved_spec.window_frame.is_some() {
-                window_frame = resolved_spec.window_frame;
-            }
-
-            // replace with new window spec
-            let new_window_spec = WindowSpec {
-                existing_window_name: None,
-                partition_by,
-                order_by,
-                window_frame,
-            };
-            Ok(new_window_spec)
-        } else {
-            Ok(window_spec.clone())
-        }
     }
 }
