+++ tests/sqllogictests/suites/query/window_function/window_in_expr.test
@@ -1,0 +1,52 @@
+statement ok
+CREATE DATABASE IF NOT EXISTS test_window_in_expr
+
+statement ok
+USE test_window_in_expr
+
+statement ok
+DROP TABLE IF EXISTS t1
+
+statement ok
+CREATE TABLE t1(a int)
+
+statement ok
+INSERT INTO t1 VALUES (1),(1),(1),(3),(3),(5),(5)
+
+query II
+select a, 1 + sum(a) over (partition by a) from t1;
+----
+1 4
+1 4
+1 4
+3 7
+3 7
+5 11
+5 11
+
+query IIII
+select 
+    a, 
+    1 + sum(a) over w, 
+    2 + sum(a) over w1, 
+    3 + sum(a) over w2 
+from t1 
+window 
+    w as (partition by a), 
+    w2 as (w1 rows current row), 
+    w1 as (w order by a) 
+order by a
+----
+1 4 5 4
+1 4 5 4
+1 4 5 4
+3 7 8 6
+3 7 8 6
+5 11 12 8
+5 11 12 8
+
+statement ok
+USE default
+
+statement ok
+DROP DATABASE test_window_in_expr
